step 1 of many to clean up the class structure: generate sensible Propel classes
This basically moves lib/core/ to db/ and makes it so that the generated classes
reside in src/Datawrapper/ORM. Also, they are now properly namespaced.

The Propel autoloading (the classmap stuff in datawrapper-conf.php) has been
disabled because we use Composer anyway. That's why the generated classmap
is now also gitignored.