use Grunt to build the dw-2.0[.min].js
This eliminates the need for a bash. Not sure if we should have this build script
in dw.js/ instead, cause it more or less looks like that directory contains a
self-contained node package, even though "dw.js" is not registred at npmjs.

As I plan to use Grunt for more things, I thought it would make sense to have one
build script for the whole datawrapper project instead.